third_party/WebKit/LayoutTests/http/tests/fetch/chromium/error-messages-expected.txt

CONSOLE ERROR: Failed to load http://localhost:8000/fetch/resources/doctype.html: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access. If an opaque response serves your needs, set the request's mode to 'no-cors' to fetch the resource with CORS disabled.
CONSOLE ERROR: Fetch API cannot load http://localhost:8000/fetch/resources/redirect-loop.php?Count=100&ACAOrigin=*. Redirect failed.
This is a testharness.js-based test.
PASS fetch() with 200 should not output error messages
PASS fetch() with 404 should not output error messages
PASS net::ERR_TOO_MANY_REDIRECTS should not output error messages
PASS net::ERR_INVALID_CHUNKED_ENCODING should not output error messages
PASS A didFailAccessControlCheck() error should output an error message
PASS A didFailRedirectCheck() error should output an error message
Harness: the test ran to completion.
